requirements.txt
# Django
Django>=4.2,<5.0

# Cryptography
pycryptodome>=3.19.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0

# Code quality (optional, for local checks)
flake8>=6.1.0
black>=23.0.0
